Reorganize pds image processing code
