+++ packages/pds/src/image/index.ts
@@ -1,0 +1,2 @@
+export * from './sharp'
+export type { ImageProcessor, Options, ImageInfo } from './util'
